requirements.txt
python-dotenv
colorama
semver
tqdm
